carrito terminado y alerts

+++ js/carrito.js
@@ -17,7 +17,7 @@ function mostrarCarrito(arrayDeProductos){
     for (const prod of arrayDeProductos){
         modalCarrito.innerHTML +=
         `
-        <div class="card mb-3" style="max-width: 540px;">
+        <div class="card mb-3" id = "productoCarrito${prod.id}" style="max-width: 540px;">
             <div class="row g-0">
                 <div class="col-md-4">
                 <img src="assets/${prod.imagen}" class="img-fluid rounded-start" alt="...">
@@ -32,16 +32,31 @@ function mostrarCarrito(arrayDeProductos){
             </div>
         </div>
         `
-        
-        let botonEliminar = document.getElementById(`botonEliminar${prod.id}`)
-        botonEliminar.addEventListener("click", () => {
-            eliminarUnProducto(prod.id)
-            console.log(`el producto ${prod.nombre} ha sido eliminado`)
-            console.log(prod.id)
-            
-        })
-
+    
     }
+    
+    for (const prod of arrayDeProductos){
+        document.getElementById(`botonEliminar${prod.id}`).addEventListener("click", () => {
+            console.log(`el producto ${prod.nombre} (${prod.id}) ha sido eliminado`)
+            let prodEnCarrito = document.getElementById(`productoCarrito${prod.id}`)
+            prodEnCarrito.remove()
+            Swal.fire({
+                title: 'Eliminado',
+                text: `El producto ${prod.nombre} ha sido eliminado`,
+                icon: 'info',
+                timer: 1500,
+                showConfirmButton: false
+            })
+            //borrar del array
+            let prodAEliminar = arrayDeProductos.find((producto) => producto.id == prod.id) // encontramos el obj a eliminar
+            let posicionProd = arrayDeProductos.indexOf(prodAEliminar)
+
+            arrayDeProductos.splice(posicionProd,1)
+            localStorage.setItem("carrito", JSON.stringify(arrayDeProductos))
+            carritoTotal(arrayDeProductos)
+        })
+    }
+    carritoTotal(arrayDeProductos)
 }
 
 //Botón "VACIAR CARRITO" => elimina todos los productos al mismo tiempo.
@@ -57,16 +72,35 @@ function vaciarCarrito(arrayDeProductos){
 
 
 function agregarProducto(producto){
-    carrito.push(producto)
-    localStorage.setItem("carrito", JSON.stringify(carrito))
+    const existe = carrito.find((elem) => elem.id == producto.id)
+    if (existe == undefined){
+
+        carrito.push(producto)
+        localStorage.setItem("carrito", JSON.stringify(carrito))
+        Swal.fire({
+            title: 'Producto Agregado!',
+            text: `El producto ${producto.nombre} ha sido agregado al carrito`,
+            icon: 'success',
+            timer: 2000,
+            showConfirmButton: false
+        })
+    } else{
+        console.log("ya existe")
+        Swal.fire({
+            title: 'Proudcto ya agregado!',
+            text: `El producto ${producto.nombre} ya existe en el carrito`,
+            icon: 'warning',
+            timer: 2000,
+            showConfirmButton: false
+        })}
 }
 
-function eliminarUnProducto(id){
-    // console.log("entró")
-    
+
+let precioTotal = document.getElementById("precioTotal")
+function carritoTotal(arrayDeProductos){
+    let total = arrayDeProductos.reduce((acc, productoCarrito)=>acc + productoCarrito.precio ,0)
+    precioTotal.innerHTML = `El total es <strong>$${total}</strong>`
 }
-
-
 //ICONO CARRITO : muestra los productos agregados al array Carrito.
 let botonCarrito = document.getElementById("botonCarrito")
